Use market top details component
