+++ app/src/components/market/market_view_container.tsx
@@ -1,37 +1,13 @@
-import { useQuery } from '@apollo/react-hooks'
-import { BigNumber } from 'ethers/utils'
-import gql from 'graphql-tag'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
 import { useMarketMakerData } from '../../hooks'
 import { useConnectedWeb3Context } from '../../hooks/connectedWeb3'
-import { getLogger } from '../../util/logger'
 import { FullLoading } from '../loading'
 
 import { MarketView } from './market_view'
 
-const logger = getLogger('Market::View')
-
 interface Props {
   marketMakerAddress: string
-}
-
-const GET_COLLATERAL_VOLUME_NOW = gql`
-  query Current($id: String) {
-    fixedProductMarketMakers(where: { id: $id }) {
-      collateralVolume
-    }
-  }
-`
-
-const buildQuery24hsEarlier = (hash: Maybe<string>) => {
-  return gql`
-  query AfterHash($id: String) {
-    fixedProductMarketMakers(where: { id: $id }, block: { hash: "${hash}" }) {
-      collateralVolume
-    }
-  }
-`
 }
 
 const MarketViewContainer: React.FC<Props> = (props: Props) => {
@@ -39,34 +15,7 @@ const MarketViewContainer: React.FC<Props> = (props: Props) => {
 
   const { marketMakerAddress } = props
 
-  const [hash, setHash] = useState<Maybe<string>>(null)
   const { marketMakerData, status } = useMarketMakerData(marketMakerAddress, context)
-  const { library: provider } = context
-
-  const [lastDayVolume, setLastDayVolume] = useState<Maybe<BigNumber>>(null)
-
-  const { data: volumeNow, error: errorVolumeNow } = useQuery(GET_COLLATERAL_VOLUME_NOW, {
-    skip: !!lastDayVolume,
-    variables: { id: marketMakerAddress.toLowerCase() },
-  })
-
-  const { data: volumeBefore, error: errorVolumeBefore } = useQuery(buildQuery24hsEarlier(hash && hash.toLowerCase()), {
-    skip: !!lastDayVolume || !hash,
-    variables: { id: marketMakerAddress.toLowerCase() },
-  })
-
-  if (errorVolumeBefore || errorVolumeNow) {
-    setLastDayVolume(null)
-    errorVolumeBefore && logger.log(errorVolumeBefore)
-    errorVolumeNow && logger.log(errorVolumeNow)
-  } else if (volumeNow && volumeBefore) {
-    const marketNow = volumeNow.fixedProductMarketMakers[0]
-    const marketBefore = volumeBefore.fixedProductMarketMakers[0]
-    const now = new BigNumber(marketNow ? marketNow.collateralVolume : 0)
-    const before = new BigNumber(marketBefore ? marketBefore.collateralVolume : 0)
-
-    setLastDayVolume(now.sub(before))
-  }
 
   const {
     arbitrator,
@@ -83,18 +32,6 @@ const MarketViewContainer: React.FC<Props> = (props: Props) => {
     userPoolShares,
   } = marketMakerData
 
-  useEffect(() => {
-    const get24hsVolume = async () => {
-      const BLOCKS_PER_SECOND = 15
-      const OFFSET = Math.round((60 * 60 * 24) / BLOCKS_PER_SECOND)
-      const lastBlock = await provider.getBlockNumber()
-      const { hash } = await provider.getBlock(lastBlock - OFFSET)
-      setHash(hash)
-    }
-
-    get24hsVolume()
-  }, [provider])
-
   if (!collateral) {
     return <FullLoading />
   }
@@ -109,7 +46,6 @@ const MarketViewContainer: React.FC<Props> = (props: Props) => {
       funding={marketMakerFunding}
       isConditionResolved={isConditionResolved}
       isQuestionFinalized={isQuestionFinalized}
-      lastDayVolume={lastDayVolume}
       marketMakerAddress={marketMakerAddress}
       question={question || ''}
       questionId={questionId}
