+++ app/src/components/market/profile/view.tsx
@@ -8,18 +8,10 @@ import { formatBigNumber } from '../../../util/tools'
 import { Arbitrator, BalanceItem, OutcomeTableValue, Status, Token } from '../../../util/types'
 import { Button, ButtonContainer } from '../../button'
 import { ButtonType } from '../../button/button_styling_types'
-import {
-  DisplayArbitrator,
-  GridTwoColumns,
-  SubsectionTitle,
-  SubsectionTitleAction,
-  SubsectionTitleWrapper,
-  ThreeBoxComments,
-  TitleValue,
-  ViewCard,
-} from '../../common'
+import { ThreeBoxComments, ViewCard } from '../../common'
 import { OutcomeTable } from '../../common/outcome_table'
 import { FullLoading } from '../../loading'
+import { MarketTopDetails } from '../market_top_details'
 
 const LeftButton = styled(Button)`
   margin-right: auto;
@@ -36,29 +28,12 @@ interface Props extends RouteComponentProps<{}> {
   status: Status
   marketMakerAddress: string
   funding: BigNumber
-  lastDayVolume: Maybe<BigNumber>
   totalPoolShares: BigNumber
   userPoolShares: BigNumber
 }
 
 const ViewWrapper = (props: Props) => {
-  const {
-    arbitrator,
-    balances,
-    category,
-    collateral,
-    history,
-    lastDayVolume,
-    marketMakerAddress,
-    questionId,
-    status,
-    totalPoolShares,
-    userPoolShares,
-  } = props
-  const [showingExtraInformation, setExtraInformation] = useState(false)
-
-  const toggleExtraInformation = () =>
-    showingExtraInformation ? setExtraInformation(false) : setExtraInformation(true)
+  const { balances, collateral, history, marketMakerAddress, status } = props
 
   const userHasShares = balances.some((balanceItem: BalanceItem) => {
     const { shares } = balanceItem
@@ -83,61 +58,14 @@ const ViewWrapper = (props: Props) => {
     )
   }
 
-  const details = (showExtraDetails: boolean) => {
-    return (
-      <>
-        <GridTwoColumns>
-          {category && <TitleValue title={'Category'} value={category} />}
-          <TitleValue title={'Arbitrator'} value={arbitrator && <DisplayArbitrator arbitrator={arbitrator} />} />
-          {questionId && (
-            <TitleValue
-              title={'Realitio'}
-              value={
-                <a
-                  href={`https://realitio.github.io/#!/question/${questionId}`}
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  Question URL
-                </a>
-              }
-            />
-          )}
-          <TitleValue
-            title="24h Volume"
-            value={lastDayVolume ? `${formatBigNumber(lastDayVolume, collateral.decimals)} ${collateral.symbol}` : '-'}
-          />
-          {showExtraDetails ? (
-            <>
-              <TitleValue
-                title={'Total Pool Tokens'}
-                value={totalPoolShares ? formatBigNumber(totalPoolShares, collateral.decimals) : '0'}
-              />
-              <TitleValue
-                title={'My Pool Tokens'}
-                value={userPoolShares ? formatBigNumber(userPoolShares, collateral.decimals) : '0'}
-              />
-              <TitleValue title={'Total Pool Earnings'} value={'Mocked Value 3'} />
-              <TitleValue title={'My Earnings'} value={'Mocked Value 4'} />
-            </>
-          ) : null}
-        </GridTwoColumns>
-      </>
-    )
-  }
-
-  const marketHasDetails = category || arbitrator
-
   return (
     <>
       <ViewCard>
-        <SubsectionTitleWrapper>
-          <SubsectionTitle>Market Information</SubsectionTitle>
-          <SubsectionTitleAction onClick={toggleExtraInformation}>
-            {showingExtraInformation ? 'Hide' : 'Show'} Pool Information
-          </SubsectionTitleAction>
-        </SubsectionTitleWrapper>
-        {marketHasDetails && details(showingExtraInformation)}
+        <MarketTopDetails
+          title="Purchase Outcome"
+          toggleTitleAction="Pool Information"
+          marketMakerAddress={marketMakerAddress}
+        />
         {renderTableData()}
         <WhenConnected>
           <ButtonContainer>
